+++ include/PgSQL_Protocol.h
@@ -284,151 +284,6 @@ private:
 	friend void SQLite3_to_Postgres(PtrSizeArray* psa, SQLite3_result* result, char* error, int affected_rows, const char* query_type, char txn_state);
 };
 
-class PgSQL_Parse_Message {
-public:
-	PgSQL_Parse_Message();
-	~PgSQL_Parse_Message();
-	const char* stmt_name = NULL;		// The name of the prepared statement
-	const char* query_string = NULL;	// The query string to be prepared
-	uint16_t num_param_types = 0;		// Number of parameter types specified
-	const uint32_t* param_types = NULL;	// Array of parameter types (can be nullptr if none)
-
-	bool parse(PtrSize_t& pkt);
-	PtrSize_t detach();
-
-private:
-	PtrSize_t _pkt = {};
-};
-
-class PgSQL_Describe_Message {
-public:
-	PgSQL_Describe_Message();
-	~PgSQL_Describe_Message();
-	uint8_t stmt_type = 0;		// 'S' for statement, 'P' for portal
-	const char* stmt_name = NULL;	// The name of the prepared statement or portal
-	/**
-	 * @brief Parses the PgSQL_Describe_Message from the provided packet.
-	 *
-	 * This method extracts the statement type and name from the packet and
-	 * initializes the internal state of the PgSQL_Describe_Message object.
-	 *
-	 * @param pkt The packet containing the PgSQL_Describe_Message data.
-	 *
-	 * @return True if parsing was successful, false otherwise.
-	 */
-	bool parse(PtrSize_t& pkt);
-	PtrSize_t detach();
-
-private:
-	PtrSize_t _pkt = {};
-};
-
-//create close statement message 
-class PgSQL_Close_Message {
-public:
-	PgSQL_Close_Message();
-	~PgSQL_Close_Message();
-	uint8_t stmt_type = 0;		// 'S' for statement, 'P' for portal
-	const char* stmt_name = NULL;	// The name of the prepared statement or portal
-	/**
-	 * @brief Parses the PgSQL_Close_Message from the provided packet.
-	 *
-	 * This method extracts the statement type and name from the packet and
-	 * initializes the internal state of the PgSQL_Close_Message object.
-	 *
-	 * @param pkt The packet containing the PgSQL_Close_Message data.
-	 *
-	 * @return True if parsing was successful, false otherwise.
-	 */
-	bool parse(PtrSize_t& pkt);
-	PtrSize_t detach();
-private:
-	PtrSize_t _pkt = {};
-};
-
-class PgSQL_Bind_Message {
-public:
-	typedef struct {
-		int32_t len;         // Length of value (-1 for NULL)
-		const unsigned char* value;  // Pointer to value data
-	} ParamValue_t;
-
-	// Iterator context for parameter values
-	typedef struct {
-		const unsigned char* current;   // Current position in values
-		uint16_t remaining;            // Parameters remaining
-	} ParamValueIterCtx;
-
-	// Iterator context for format arrays
-	typedef struct {
-		const unsigned char* current;   // Current position in array
-		uint16_t remaining;         // Formats remaining
-	} FormatIterCtx;
-
-	PgSQL_Bind_Message();
-	~PgSQL_Bind_Message();
-	const char* portal_name = NULL;		// The name of the portal to bind
-	const char* stmt_name = NULL;		// The name of the prepared statement to bind
-	uint16_t num_param_formats = 0;		// Number of parameters to bind
-	uint16_t num_param_values = 0;
-	uint16_t num_result_formats = 0; // Number of result format codes
-	/**
-	 * @brief Parses the PgSQL_Bind_Message from the provided packet.
-	 *
-	 * This method extracts the portal name, statement name, parameter formats,
-	 * parameter values, and result formats from the packet and initializes the
-	 * internal state of the PgSQL_Bind_Message object.
-	 *
-	 * @param pkt The packet containing the PgSQL_Bind_Message data.
-	 *
-	 * @return True if parsing was successful, false otherwise.
-	 */
-	bool parse(PtrSize_t& pkt);
-	PtrSize_t detach();
-	PgSQL_Bind_Message* release();
-
-	// Initialize param format iterator
-	void init_param_format_iter(FormatIterCtx* ctx) const;
-	// Initialize parameter value iterator
-	void init_param_value_iter(ParamValueIterCtx* ctx) const;
-	// Get next parameter value
-	bool next_param_value(ParamValueIterCtx* ctx, ParamValue_t* out) const;
-	// Initialize result format iterator
-	void init_result_format_iter(FormatIterCtx* ctx) const;
-	// Get next format value
-	bool next_format(FormatIterCtx* ctx, uint16_t* out) const;
-
-private:
-	const uint16_t* param_formats = NULL;	// Array of parameter types (can be nullptr if none)
-	const uint8_t* param_values = NULL;		// Array of parameter values (can be nullptr if none)
-	const uint16_t* result_formats = NULL;		// Array of result format codes
-
-	PtrSize_t _pkt = {};
-};
-
-class PgSQL_Execute_Message {
-public:
-	PgSQL_Execute_Message();
-	~PgSQL_Execute_Message();
-	const char* portal_name = NULL;		// The name of the portal to execute
-	uint32_t max_rows = 0;				// Maximum number of rows to return (0 for no limit)
-	/**
-	 * @brief Parses the PgSQL_Execute_Message from the provided packet.
-	 *
-	 * This method extracts the portal name and maximum rows from the packet
-	 * and initializes the internal state of the PgSQL_Execute_Message object.
-	 *
-	 * @param pkt The packet containing the PgSQL_Execute_Message data.
-	 *
-	 * @return True if parsing was successful, false otherwise.
-	 */
-	bool parse(PtrSize_t& pkt);
-	PtrSize_t detach();
-
-private:
-	PtrSize_t _pkt = {};
-};
-
 class PgSQL_Protocol;
 
 struct ColumnMetadata {
